Wrapper/requirements.txt
python-dotenv
sentry-sdk

langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langgraph

pypdf

faiss-cpu
sentence-transformers

langsmith